# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API PERICLES",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Show all appointments saved",
                "description": "Show all appointments saved in the database",
                "operationId": "158f8acc8087367fcc7bc952790c9473",
                "responses": {
                    "200": {
                        "description": "ALL APPOINTMENTS"
                    },
                    "404": {
                        "description": "ERROR"
                    }
                }
            }
        },
        "/api/appointment/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get Appointment By ID",
                "description": "Find appointment by id.",
                "operationId": "ad059938c18d2b24d9d9d3ede9d1194f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Get Appointment By Id ",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Appointment By Id"
                    },
                    "404": {
                        "description": "ERROR"
                    }
                }
            }
        },
        "/api/appointment/dni/client/{DNI_Client}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get Appointment By DNI_Client",
                "description": "Find appointment by DNI_Client",
                "operationId": "bcb28d9e9fedf7ee0e272240e40dd18b",
                "parameters": [
                    {
                        "name": "DNI_Client",
                        "in": "query",
                        "description": "Get Appointment By DNI_Client",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Appointment By DNI_Client"
                    },
                    "404": {
                        "description": "ERROR"
                    }
                }
            }
        },
        "/api/appointment/dni/student/{DNI_Student}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get Appointment By DNI_Student",
                "description": "Find appointment by DNI_Student",
                "operationId": "dbed6b250717e4ca30df53a866bda75f",
                "parameters": [
                    {
                        "name": "DNI_Student",
                        "in": "query",
                        "description": "Get Appointment By DNI_Student",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Appointment By DNI_Student"
                    },
                    "404": {
                        "description": "ERROR"
                    }
                }
            }
        },
        "/api/appointment/add": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Add appointment",
                "description": "Add appointment",
                "operationId": "d3be3893cfcd9b24812f509ab841bbfe",
                "responses": {
                    "200": {
                        "description": "Quote added"
                    },
                    "404": {
                        "description": "ERROR"
                    }
                }
            }
        },
        "/api/appointment/edit/{id}": {
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Edit appointment by id",
                "description": "Edit appointment by id",
                "operationId": "22262a9ce778cebdb5f9cf5b42b03d1b",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "query",
                        "description": "Edit appointment by id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edited quote"
                    },
                    "404": {
                        "description": "ERROR"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete appointment by id",
                "description": "Delete appointment by id",
                "operationId": "dcaee28126a241eabe335fb87a229b57",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "query",
                        "description": "Delete appointment by id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted date"
                    },
                    "404": {
                        "description": "ERROR"
                    }
                }
            }
        },
...
    }
}